add middlewares

+++ internal/api/handlers/article_handler.go
@@ -2,35 +2,23 @@ package handlers
 
 import (
 	"edjr-trk/internal/api/dto"
-	"edjr-trk/internal/api/validators" // Подключение кастомного валидатора
 	"edjr-trk/internal/service"
 	"edjr-trk/pkg/http_error"
-	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
 	"go.uber.org/zap"
-	"strconv"
 )
 
 type ArticleHandler struct {
-	service  *service.ArticleService
-	logger   *zap.Logger
-	validate *validator.Validate
+	service *service.ArticleService
+	logger  *zap.Logger
 }
 
 // NewArticleHandler creates a new instance of ArticleHandler.
 func NewArticleHandler(service *service.ArticleService, logger *zap.Logger) *ArticleHandler {
-	validate := validator.New()
-	validators.RegArticleValidators(validate) // Регистрируем кастомные валидаторы
 	return &ArticleHandler{
-		service:  service,
-		logger:   logger,
-		validate: validate,
+		service: service,
+		logger:  logger,
 	}
-}
-
-// ValidateStruct validates the given input structure.
-func (h *ArticleHandler) ValidateStruct(input interface{}) error {
-	return h.validate.Struct(input)
 }
 
 // CreateArticle handles creating a new article.
@@ -59,17 +47,19 @@ func (h *ArticleHandler) CreateArticle(c *fiber.Ctx) error {
 func (h *ArticleHandler) GetAllArticles(c *fiber.Ctx) error {
 	h.logger.Info("Received request to fetch paginated articles")
 
-	// Parse the page number.
-	pageStr := c.Query("page", "1")
-	pageNumber, err := strconv.Atoi(pageStr)
-	if err != nil || pageNumber < 1 {
+	// Retrieve pagination parameters from context.
+	pageNumberInterface := c.Locals("pageNumber")
+	pageSizeInterface := c.Locals("pageSize")
+
+	pageNumber, ok := pageNumberInterface.(int)
+	if !ok {
+		h.logger.Error("Failed to retrieve page number from context")
 		pageNumber = 1
 	}
 
-	// Parse the page size.
-	sizeStr := c.Query("size", "10")
-	pageSize, err := strconv.Atoi(sizeStr)
-	if err != nil || pageSize < 1 {
+	pageSize, ok := pageSizeInterface.(int)
+	if !ok {
+		h.logger.Error("Failed to retrieve page size from context")
 		pageSize = 10
 	}
 
@@ -89,46 +79,22 @@ func (h *ArticleHandler) GetAllArticles(c *fiber.Ctx) error {
 	return c.Status(fiber.StatusOK).JSON(articles)
 }
 
-// formatValidationErrors formats validation errors into an array of ErrorItem.
-func (h *ArticleHandler) formatValidationErrors(validationErrors validator.ValidationErrors) []http_error.ErrorItem {
-	var errorDetails []http_error.ErrorItem
-	for _, validationErr := range validationErrors {
-		field := validationErr.Field() // The field that failed validation
-		tag := validationErr.Tag()     // The validation tag that failed
-		var errorMsg string
-		switch tag {
-		case "required":
-			errorMsg = "This field is required"
-		case "min":
-			errorMsg = "The field does not meet the minimum length requirement"
-		case "img_base64_or_null":
-			errorMsg = "The field must be null or a valid Base64 string"
-		default:
-			errorMsg = "Validation failed on tag: " + tag
-		}
-		errorDetails = append(errorDetails, http_error.ErrorItem{
-			Field: field,
-			Error: errorMsg,
-		})
-	}
-	return errorDetails
-}
-
 // GetArticleById handles fetching a single article by its ID.
 func (h *ArticleHandler) GetArticleById(c *fiber.Ctx) error {
 	h.logger.Info("Received request to fetch an article by ID")
 
-	// Получаем ID статьи из параметров маршрута.
-	articleId := c.Params("id")
-	if articleId == "" {
-		h.logger.Error("Article ID is missing in the request")
-		return http_error.NewHTTPError(fiber.StatusBadRequest, "Article ID is required", nil).Send(c)
+	// Retrieve article ID from context.
+	articleIDInterface := c.Locals("articleID")
+	articleID, ok := articleIDInterface.(string)
+	if !ok || articleID == "" {
+		h.logger.Error("Article ID is missing in the context")
+		return http_error.NewHTTPError(fiber.StatusInternalServerError, "Article ID is required", nil).Send(c)
 	}
 
-	// Получаем статью через сервис.
-	article, err := h.service.GetArticleById(c.Context(), articleId)
+	// Fetch the article via the service.
+	article, err := h.service.GetArticleById(c.Context(), articleID)
 	if err != nil {
-		h.logger.Error("Failed to fetch article by ID", zap.String("articleId", articleId), zap.Error(err))
+		h.logger.Error("Failed to fetch article by ID", zap.String("articleID", articleID), zap.Error(err))
 		return http_error.NewHTTPError(fiber.StatusInternalServerError, "Failed to fetch article", nil).Send(c)
 	}
 
@@ -136,29 +102,30 @@ func (h *ArticleHandler) GetArticleById(c *fiber.Ctx) error {
 		return http_error.NewHTTPError(fiber.StatusNotFound, "Article not found", nil).Send(c)
 	}
 
-	h.logger.Info("Article fetched successfully", zap.String("articleId", articleId))
+	h.logger.Info("Article fetched successfully", zap.String("articleID", articleID))
 	return c.Status(fiber.StatusOK).JSON(article)
 }
 
-// RemoveArticleById handles partial updates of an article.
+// RemoveArticleById handles removing an article by its ID.
 func (h *ArticleHandler) RemoveArticleById(c *fiber.Ctx) error {
 	h.logger.Info("Received request to remove an article")
 
-	// Получаем ID статьи из параметров маршрута.
-	articleID := c.Params("id")
-	if articleID == "" {
-		h.logger.Error("Article ID is missing in the request")
-		return http_error.NewHTTPError(fiber.StatusBadRequest, "Article ID is required", nil).Send(c)
+	// Retrieve article ID from context.
+	articleIDInterface := c.Locals("articleID")
+	articleID, ok := articleIDInterface.(string)
+	if !ok || articleID == "" {
+		h.logger.Error("Article ID is missing in the context")
+		return http_error.NewHTTPError(fiber.StatusInternalServerError, "Article ID is required", nil).Send(c)
 	}
 
-	// Обновляем статью через сервис.
+	// Remove the article via the service.
 	removedArticle, err := h.service.RemoveArticleById(c.Context(), articleID)
 	if err != nil {
 		h.logger.Error("Failed to remove article", zap.String("articleID", articleID), zap.Error(err))
 		return http_error.NewHTTPError(fiber.StatusInternalServerError, "Failed to remove article", nil).Send(c)
 	}
 
-	h.logger.Info("Article remove successfully", zap.String("articleID", articleID))
+	h.logger.Info("Article removed successfully", zap.String("articleID", articleID))
 	return c.Status(fiber.StatusOK).JSON(fiber.Map{"id": removedArticle})
 }
 
@@ -166,35 +133,23 @@ func (h *ArticleHandler) RemoveArticleById(c *fiber.Ctx) error {
 func (h *ArticleHandler) PatchArticleById(c *fiber.Ctx) error {
 	h.logger.Info("Received request to patch an article")
 
-	// Получаем ID статьи из параметров маршрута.
-	articleID := c.Params("id")
-	if articleID == "" {
-		h.logger.Error("Article ID is missing in the request")
-		return http_error.NewHTTPError(fiber.StatusBadRequest, "Article ID is required", nil).Send(c)
+	// Retrieve article ID from context.
+	articleIDInterface := c.Locals("articleID")
+	articleID, ok := articleIDInterface.(string)
+	if !ok || articleID == "" {
+		h.logger.Error("Article ID is missing in the context")
+		return http_error.NewHTTPError(fiber.StatusInternalServerError, "Article ID is required", nil).Send(c)
 	}
 
-	// Парсим входящие данные.
-	var req dto.PatchArticleRequest
-	if err := c.BodyParser(&req); err != nil {
-		h.logger.Error("Failed to parse request body", zap.Error(err))
-		return http_error.NewHTTPError(fiber.StatusBadRequest, "Invalid request body", nil).Send(c)
+	// Retrieve validated data from context.
+	reqInterface := c.Locals("validatedBody")
+	req, ok := reqInterface.(dto.PatchArticleRequest)
+	if !ok {
+		h.logger.Error("Failed to retrieve validated request from context")
+		return http_error.NewHTTPError(fiber.StatusInternalServerError, "Internal Server Error", nil).Send(c)
 	}
 
-	// Validate the input data.
-	if err := h.ValidateStruct(&req); err != nil {
-		h.logger.Error("Validation failed for request body", zap.Error(err))
-
-		// If the error is a validation error.
-		if validationErrors, ok := err.(validator.ValidationErrors); ok {
-			errorDetails := h.formatValidationErrors(validationErrors)
-			return http_error.NewHTTPError(fiber.StatusBadRequest, "Validation error", errorDetails).Send(c)
-		}
-
-		// For other validation errors.
-		return http_error.NewHTTPError(fiber.StatusBadRequest, "Invalid input", nil).Send(c)
-	}
-
-	// Обновляем статью через сервис.
+	// Update the article via the service.
 	updatedArticle, err := h.service.PatchArticleById(c.Context(), req, articleID)
 	if err != nil {
 		h.logger.Error("Failed to update article", zap.String("articleID", articleID), zap.Error(err))
